[ESI][Runtime] Make esitester public
- Expose the esitester pycde script in a distributed module.
- Distribute the esitester software binary.
